Refactored networking into NetworkController
